Create auth.sh
This is a script meant to live inside of the provisioned Docker container to demonstrate that the secret credentials were successfully passed. The script requires that cURL be installed (on debian use apt-get install curl)

The paypayRegistrationId is expected to have expired, but this will still demonstrate a successful connection to the vantiv eCommerce endpoint using the passed credentials.